Stop the iterative tests. A renew tests.
