+++ certbot-ci/certbot_integration_tests/utils/assertions.py
@@ -12,6 +12,7 @@ def assert_hook_execution(probe_path, probe_content):
 def assert_save_renew_hook(config_dir, lineage):
     assert os.path.isfile(os.path.join(config_dir, 'renewal/{0}.conf'.format(lineage)))
 
+
 def assert_certs_count_for_lineage(config_dir, lineage, count):
     archive_dir = os.path.join(config_dir, 'archive')
     lineage_dir = os.path.join(archive_dir, lineage)
