+++ certbot-ci/certbot_integration_tests/utils/fixtures_session.py
@@ -1,8 +1,4 @@
 import os
-import subprocess
-import tempfile
-import shutil
-import sys
 import pytest
 
 
@@ -18,26 +14,6 @@ def acme_url():
         return 'https://localhost:14000/dir'
 
     raise ValueError('Invalid CERTBOT_INTEGRATION value: {0}'.format(integration))
-
-
-@pytest.fixture(scope='session')
-def workspace():
-    workspace = tempfile.mkdtemp()
-    try:
-        yield workspace
-    finally:
-        shutil.rmtree(workspace)
-
-
-@pytest.fixture(scope='session')
-def config_dir(workspace):
-    return os.path.join(workspace, 'conf')
-
-
-@pytest.fixture(scope='session')
-def renewal_hooks_dirs(config_dir):
-    renewal_hooks_root = os.path.join(config_dir, 'renewal-hooks')
-    return [os.path.join(renewal_hooks_root, item) for item in ['pre', 'deploy', 'post']]
 
 
 @pytest.fixture(scope='session')
